Mathematical Functions General Project
The primary purpose of this program is to create a base class that can generate random numbers as requested by the programmer or user, returning them either as integers or as floating-point numbers (between 0 and 1), so that they can be used for various purposes.

+++ RandomNumbers.cpp
@@ -371,7 +371,7 @@ int main()
 
 				for (int ind = V_ZERO; ind < numbers; ind++)
 					{
-						ticket.insert((*array_Random_Number[idx]).getWithin(minimum, maximum));
+						ticket.insert(static_cast<int>((*array_Random_Number[idx]).getWithin(minimum, maximum)));
 
 						std::cout << "#: [" << (counter++) + V_ONE << "]\t:\t[" << array_Random_Number[idx]->getNumber() << "]\t=\t[" << array_Random_Number[idx]->getValue() << "]\t=\t[" << array_Random_Number[idx]->getWithin(minimum, maximum) << "]." << std::endl;
 						(*array_Random_Number[idx])();	//Activate the generator to get the next number.
